Add files via upload
The central question guiding this project is: "What if all the trash in our world became 
alive? How could we stop it?" This hypothetical scenario serves as the foundation for the game's 
narrative, making the concept of waste management more relatable and engaging for players. 
"Litter Legends" addresses this question by transforming trash collection and recycling into a fun 
and educational activity. Players navigate through various levels, each presenting different types 
of trash and environmental challenges. As they progress, they learn about the environmental 
impact of different waste items and receive tips on proper disposal methods. The game aims to 
create a sense of urgency and responsibility in players, motivating them to act against 
environmental pollution.
 By blending an immersive storyline, engaging gameplay mechanics, and educational 
content, "Litter Legends" aims to inspire players to become real-life heroes in the fight against 
pollution and waste mismanagement. Through this entertaining and informative experience, 
players develop a deeper understanding of the importance of proper waste disposal and recycling, 
empowering them to make positive changes in their daily lives and contribute to a cleaner, more 
sustainable future